requirement.txt
gradio
numpy==1.24.4
scikit-learn
pillow
joblib
